+++ client-app/src/main/java/com/piedraazul/app_client/services/ProfessionalService.java
@@ -8,7 +8,7 @@ import com.piedraazul.app_client.models.Professional;
 public interface ProfessionalService {
     List<Professional> getAllProfessionals();
 
-    Professional findByCod(int codProf);
+    Professional findByCod(Long codProf);
 
     List<Professional> getAllProfessionalsBySpeciality(SpecialityProfEnum speciality);
